requirements.txt
docling
lxml_html_clean
beautifulsoup4
trafilatura
langdetect
transformers
protobuf
sentencepiece
qdrant-client
torch
llama-cpp-python
sentence-transformers
tiktoken
psycopg[binary]
spacy
mutagen
odfpy
